# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "HyperHive",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "hyperhive.vajid.tech",
    "basePath": "/",
    "paths": {
...
        "/admin/block": {
            "patch": {
                "security": [
                    {
                        "AdminAutherisation": []
                    }
                ],
                "description": "Block a user's account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Block User Account",
                "parameters": [
                    {
                        "description": "Request body for blocking user account",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel_auth_svc.UserIDReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User account blocked successfully",
                        "schema": {
                            "$ref": "#/definitions/response_auth_svc.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response_auth_svc.Response"
                        }
                    }
                }
            }
        },
        "/admin/login": {
            "post": {
                "description": "Authenticate admin and generate access token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Login",
                "parameters": [
                    {
                        "description": "Request body for admin login",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel_auth_svc.AdminLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token response",
                        "schema": {
                            "$ref": "#/definitions/response_auth_svc.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response_auth_svc.Response"
                        }
                    }
                }
            }
        },
        "/admin/unblock": {
            "patch": {
                "security": [
                    {
                        "AdminAuthorization": []
                    }
                ],
                "description": "Unblock a user's account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Unblock User Account",
                "parameters": [
                    {
                        "description": "Request body for unblocking user account",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel_auth_svc.UserIDReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User account unblocked successfully",
                        "schema": {
                            "$ref": "#/definitions/response_auth_svc.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response_auth_svc.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "requestmodel_auth_svc.AdminLogin": {
            "type": "object",
            "required": [
                "Email",
                "Password"
            ],
            "properties": {
                "Email": {
                    "type": "string"
                },
                "Password": {
                    "type": "string",
                    "minLength": 5
                }
            }
        },
...
        "requestmodel_auth_svc.UserIDReq": {
            "type": "object",
            "properties": {
                "UserID": {
                    "type": "string"
                }
            }
        },
...
        "response_auth_svc.Response": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                },
                "result": {},
                "status_code": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAutherisation": {
            "type": "apiKey",
            "name": "AccessToken",
            "in": "header"
        },
        "UserAuthorization": {
            "type": "apiKey",
            "name": "AccessToken",
            "in": "header"
        },
        "UserConfirmToken": {
            "type": "apiKey",
            "name": "ConfirmToken",
            "in": "header"
        }
    }
}